% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_list.R
\name{comp_list}
\alias{comp_list}
\title{comp_list}
\usage{
comp_list(data1, data2)
}
\arguments{
\item{data1}{List.}

\item{data2}{List.}
}
\value{
A list containing:

- $same.length: logical. Are number of elements identical?

- $length: number of elements in the 2 datasets (NULL otherwise).

- $same.names: logical. Are element names identical ?

- $name: name of elements of the 2 datasets if identical (NULL otherwise).

- $any.id.name: logical. Is there any element names identical ?

- $same.names.pos1: positions, in data1, of the element names identical in data2.

- $same.names.pos2: positions, in data2, of the compartment names identical in data1.

- $any.id.compartment: logical. is there any identical compartments ?

- $same.compartment.pos1: positions, in data1, of the compartments identical in data2.

- $same.compartment.pos2: positions, in data2, of the compartments identical in data1.

- $identical.object: logical. Are objects identical (kind of object, compartment names and content)?

- $identical.content: logical. Are content objects identical (identical compartments excluding compartment names)?
}
\description{
Compare two lists. Check and report in a list if the 2 datasets have:

- same length

- common names

- common compartments
}
\examples{
obs1 = list(a = 1:5, b = LETTERS[1:2], d = matrix(1:6)) ; 
obs2 = list(a = 1:5, b = LETTERS[1:2], d = matrix(1:6)) ; 
comp_list(obs1, obs2)


obs1 = list(1:5, LETTERS[1:2]) ; 
obs2 = list(a = 1:5, b = LETTERS[1:2]) ; 
comp_list(obs1, obs2)


obs1 = list(b = 1:5, c = LETTERS[1:2]) ; 
obs2 = list(a = 1:5, b = LETTERS[1:2], d = matrix(1:6)) ; 
comp_list(obs1, obs2)


obs1 = list(b = 1:5, c = LETTERS[1:2]) ; 
obs2 = list(LETTERS[5:9], matrix(1:6), 1:5) ; 
comp_list(obs1, obs2)
}
