% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_remod.R
\name{df_remod}
\alias{df_remod}
\title{df_remod}
\usage{
df_remod(
  data,
  quanti.col.name = "quanti",
  quali.col.name = "quali",
  safer_check = TRUE
)
}
\arguments{
\item{data}{Data frame to convert.}

\item{quanti.col.name}{Single character string. Optional name for the quanti column of the new data frame.}

\item{quali.col.name}{Single character string. Optional name for the quali column of the new data frame.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
The modified data frame.
}
\description{
If the data frame is made of n numeric columns, a new data frame is created, with the 1st column gathering all the numeric values, and the 2nd column being the name of the columns of the initial data frame. If row names were present in the initial data frame, then a new ini_rowname column is added with the names of the rows.


If the data frame is made of one numeric column and one character or factor column, a new data frame is created, with the new columns corresponding to the split numeric values (according to the character column). NA are added a the end of each column to have the same number of rows. BEWARE: in such data frame, rows are not individuals. This means that in the example below, values 10 and 20 are associated on the same row but that means nothing in term of association.
}
\examples{
obs <- data.frame(col1 = (1:4)*10, col2 = 5:8, stringsAsFactors = TRUE) ; 
rownames(obs) <- paste0("row", 1:4) ; 
obs ;
df_remod(obs, quanti.col.name = "quanti", quali.col.name = "quali")
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
