% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.safer_backone_check}
\alias{.safer_backone_check}
\title{.safer_backone_check}
\usage{
.safer_backone_check(
  x,
  na.rm,
  finite,
  safer_check,
  arg.user.setting,
  external.function.name,
  external.package.name
)
}
\arguments{
\item{x}{Numeric or logical vector or matrix or numeric table where find the initial values to calculate.}

\item{na.rm}{Single logical value. Should missing values (NA and NaN) be removed ?}

\item{finite}{Single logical value. Should infinite values (Inf and -Inf) be removed ? Warning: this argument does not remove NA and NaN. Please use the na.rm argument.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)?}

\item{arg.user.setting}{List of the argument user settings.}

\item{external.function.name}{Name of the function using the .safer_backone_check() function.}

\item{external.package.name}{Name of the package of the function using the .pack_and_function_check() function.}
}
\value{
An error message if the type of input is not correct, or input missed, nothing otherwise.
}
\description{
Check if 1) the types of input are correct and 2) an input is missed.
}
\details{
WARNING: arguments of the .safer_backone_check() function are not checked, so use carefully inside other functions
}
\examples{
\dontrun{
# Example that shouldn't be run because this is an internal function
.safer_backone_check(x = 1:3, na.rm = TRUE)
.safer_backone_check(x = c(Inf, NA), na.rm = TRUE, finite = TRUE)
}
}
\keyword{internal}
