% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min.R
\name{min}
\alias{min}
\title{min}
\usage{
min(x, na.rm = FALSE, finite = FALSE)
}
\arguments{
\item{x}{Numeric or logical vector or matrix or numeric table where find the initial values to use.}

\item{na.rm}{Single logical value. Should missing values (NA and NaN) be removed ?}

\item{finite}{Single logical value. Should infinite values (Inf and -Inf) be removed ? Warning: this argument does not remove NA and NaN. Please use the na.rm argument.}
}
\value{
The minimum value of the given arguments.
}
\description{
Find and return the minimum value of a set of data.
}
\details{
REQUIRED PACKAGES

none


REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

none
}
\examples{
vec <- c(1:3) ; min(x = vec)

vec <- c(1, 3, 5, TRUE, FALSE) ; min(x = vec)


# This example returns an error because of the character in the vector
# vec <- c(1,3,5,TRUE,"apple") ; min(x = vec)
}
