% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{round2}
\usage{
round2(data, dec.nb = 2, after.lead.zero = TRUE, safer_check = TRUE)
}
\arguments{
\item{data}{A vector of numbers (numeric or character mode).}

\item{dec.nb}{Single numeric value. Number of required decimal digits.}

\item{after.lead.zero}{Single logical value. If FALSE, rounding is performed for all the decimal numbers, whatever the leading zeros (e.g., 0.123 -> 0.12 and 0.00128 -> 0.00). If TRUE, dec.nb are taken after the leading zeros (e.g., 0.123 -> 0.12 and 0.00128 -> 0.0013).}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
The modified vector.
}
\description{
Round a vector of values, if decimal, with the desired number of decimal digits after the decimal leading zeros.
}
\details{
WARNINGS

Work well with numbers as character strings, but not always with numerical numbers because of the floating point.

Numeric values are really truncated from a part of their decimal digits, whatever options(digits) settings.

See ?.Machine or https://stackoverflow.com/questions/5173692/how-to-return-number-of-decimal-places-in-r, with the interexting formula: abs(x - round(x)) > .Machine$double.eps^0.5.
}
\examples{
ini.options <- options()$digits ; 
options(digits = 8) ; 
cat(round2(data = c(NA, 10, 100.001, 333.0001254, 12312.1235), dec.nb = 2, after.lead.zero = FALSE), "\n\n") ; 
options(digits = ini.options)
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
