% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum2.R
\name{sum2}
\alias{sum2}
\title{sum2}
\usage{
sum2(x, na.rm = FALSE, finite = FALSE, safer_check = TRUE)
}
\arguments{
\item{x}{Numeric or logical vector or matrix or numeric table where find the initial values to calculate.}

\item{na.rm}{Single logical value. Should missing values (NA and NaN) be removed ?}

\item{finite}{Single logical value. Should infinite values (Inf and -Inf) be removed ? Warning: this argument does not remove NA and NaN. Please use the na.rm argument.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
The sum of the given arguments. Returns NA if the input contains NA and the argument na.rm = TRUE, else returns a value.
}
\description{
Calculate the sum of numeric and logical value.
}
\examples{
vec <- c(1:3) ; sum2(x = vec)

vec <- c(1,3,5,TRUE) ; sum2(x = vec)



# This example returns an error because of the character in the vector
# vec <- c(1,3,5,TRUE,"apple") ; sum2(x = vec)
}
\seealso{
\code{\link{sum}}
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
