% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail2.R
\name{tail2}
\alias{tail2}
\title{tail2}
\usage{
tail2(data1, n = 6, side = "l")
}
\arguments{
\item{data1}{Any object but more dedicated for matrix, data frame or table.}

\item{n}{As in utils::tail() but for for matrix, data frame or table, number of dimension to print (10 means 10 rows and columns).}

\item{side}{Either "l" or "r" for the left or right side of the 2D object (only for matrix, data frame or table).}
}
\value{
The tail.
}
\description{
As utils::tail() but display the left or right head of big 2D objects.
}
\details{
REQUIRED PACKAGES

cuteDev


REQUIRED FUNCTIONS FROM THE cute PACKAGE

arg_check()


BEWARE

Other arguments of tail() not used.
}
\examples{
obs1 = matrix(1:30, ncol = 5, dimnames = list(letters[1:6], LETTERS[1:5])) ; 
obs1 ; 
tail2(obs1, 3, "r")
}
