% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_1d.R
\name{comp_1d}
\alias{comp_1d}
\title{comp_1d}
\usage{
comp_1d(data1, data2)
}
\arguments{
\item{data1}{Vector or factor or 1D table, or 1D matrix or 1D array.}

\item{data2}{Vector or factor or 1D table, or 1D matrix or 1D array.}
}
\description{
Compare two 1D datasets (vector or factor or 1D table, or 1D matrix or 1D array) of the same class or not. Check and report in a list if the 2 datasets have:
- same class
- common elements
- common element names (except factors)
- common levels (factors only)
}
\details{
REQUIRED PACKAGES

none


REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

none
}
\examples{
obs1 = 1:5 ; obs2 = 1:5 ; names(obs1) <- LETTERS[1:5] ; names(obs2) <- LETTERS[1:5] ; comp_1d(obs1, obs2)
obs1 = 1:5 ; obs2 = 1:5 ; names(obs1) <- LETTERS[1:5] ; comp_1d(obs1, obs2)
obs1 = 1:5 ; obs2 = 3:6 ; names(obs1) <- LETTERS[1:5] ; names(obs2) <- LETTERS[1:4] ; comp_1d(obs1, obs2)
obs1 = factor(LETTERS[1:5]) ; obs2 = factor(LETTERS[1:5]) ; comp_1d(obs1, obs2)
obs1 = factor(LETTERS[1:5]) ; obs2 = factor(LETTERS[10:11]) ; comp_1d(obs1, obs2)
obs1 = factor(LETTERS[1:5]) ; obs2 = factor(LETTERS[4:7]) ; comp_1d(obs1, obs2)
obs1 = factor(c(LETTERS[1:4], "E")) ; obs2 = factor(c(LETTERS[1:4], "F")) ; comp_1d(obs1, obs2)
obs1 = 1:5 ; obs2 = factor(LETTERS[1:5]) ; comp_1d(obs1, obs2)
obs1 = 1:5 ; obs2 = 1.1:6.1 ; comp_1d(obs1, obs2)
obs1 = as.table(1:5); obs2 = as.table(1:5) ; comp_1d(obs1, obs2)
obs1 = as.table(1:5); obs2 = 1:5 ; comp_1d(obs1, obs2)
}
