% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_list.R
\name{comp_list}
\alias{comp_list}
\title{comp_list}
\usage{
comp_list(data1, data2, safer_check = TRUE)
}
\arguments{
\item{data1}{List.}

\item{data2}{List.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
A list containing:

- $same.length: logical. Are number of elements identical?

- $length: number of elements in the 2 datasets (NULL otherwise).

- $same.names: logical. Are element names identical ?

- $name: name of elements of the 2 datasets if identical (NULL otherwise).

- $any.id.name: logical. Is there any element names identical ?

- $same.names.pos1: positions, in data1, of the element names identical in data2.

- $same.names.pos2: positions, in data2, of the compartment names identical in data1.

- $any.id.compartment: logical. is there any identical compartments ?

- $same.compartment.pos1: positions, in data1, of the compartments identical in data2.

- $same.compartment.pos2: positions, in data2, of the compartments identical in data1.

- $identical.object: logical. Are objects identical (kind of object, compartment names and content)?

- $identical.content: logical. Are content objects identical (identical compartments excluding compartment names)?
}
\description{
Compare two lists. Check and report in a list if the 2 datasets have:

- same length

- common names

- common compartments
}
\examples{
obs1 = list(1:5, LETTERS[1:2]) ; 
obs2 = list(a = 1:5, b = LETTERS[1:2]) ; 
comp_list(obs1, obs2)
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
