% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head2.R
\name{head2}
\alias{head2}
\title{head2}
\usage{
head2(data1, n = 6, side = "l", safer_check = TRUE)
}
\arguments{
\item{data1}{Any object but more dedicated for matrix, data frame or table.}

\item{n}{As in head2() but for for matrix, data frame or table, number of dimension to print (10 means 10 rows and columns).}

\item{side}{Either "l" or "r" for the left or right side of the 2D object (only for matrix, data frame or table).}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
The head.
}
\description{
As utils::head() but display the left or right head of big 2D objects.
}
\details{
BEWARE

Other arguments of head() not used.
}
\examples{
obs1 = matrix(1:30, ncol = 5, dimnames = list(letters[1:6], LETTERS[1:5])) ; 
obs1 ; 
head2(obs1, 3)
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
