% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backbone.R
\name{.arguments_check}
\alias{.arguments_check}
\title{.arguments_check}
\usage{
.arguments_check(x, na.rm = FALSE, finite = FALSE)
}
\arguments{
\item{x}{Numeric or logical vector or matrix or numeric table where find the initial values to calculate.}

\item{na.rm}{Single logical value. Should missing values (NA and NaN) be removed ?}

\item{finite}{Single logical value. Should infinite values (Inf and -Inf) be removed ? Warning: this argument does not remove NA and NaN. Please use the na.rm argument.}
}
\value{
An error message if the type of input is not correct, or input missed, nothing otherwise.
}
\description{
Check if 1) the types of input are correct and 2) an input is missed.
}
\keyword{internal}
