% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.arguments_check}
\alias{.arguments_check}
\title{.arguments_check}
\usage{
.arguments_check(x, na.rm = FALSE, finite = FALSE, external.function.name)
}
\arguments{
\item{x}{Numeric or logical vector or matrix or numeric table where find the initial values to calculate.}

\item{na.rm}{Single logical value. Should missing values (NA and NaN) be removed ?}

\item{finite}{Single logical value. Should infinite values (Inf and -Inf) be removed ? Warning: this argument does not remove NA and NaN. Please use the na.rm argument.}

\item{external.function.name}{Name of the function using the .arguments_check() function.}
}
\value{
An error message if the type of input is not correct, or input missed, nothing otherwise.
}
\description{
Check if 1) the types of input are correct and 2) an input is missed.
}
\examples{
\dontrun{
# Example that shouldn't be run because this is an internal function
.arguments_check(x = 1:3, na.rm = TRUE)

.arguments_check(x = c(Inf, NA), na.rm = TRUE, finite = TRUE)
}
}
\keyword{internal}
