% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max2.R
\name{max2}
\alias{max2}
\title{max2}
\usage{
max2(x, na.rm = FALSE, finite = FALSE, safer_check = TRUE)
}
\arguments{
\item{x}{Numeric or logical vector or matrix or numeric table where find the initial values to use.}

\item{na.rm}{Single logical value. Should missing values (NA and NaN) be removed ?}

\item{finite}{Single logical value. Should infinite values (Inf and -Inf) be removed ? Warning: this argument does not remove NA and NaN. Please use the na.rm argument.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Must be set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
The maximum value of the given arguments.
}
\description{
Find and return the maximum value of a set of data.
}
\examples{
vec <- c(1:3) ; max2(x = vec)

vec <- c(1,3,5,TRUE, FALSE, NA) ; max2(x = vec, na.rm = TRUE)


# This example returns an error because of the character in the vector
# vec <- c(1,3,5,TRUE,"apple") ; max2(x = vec)

# This example returns an error because of the argument na.rm, when the function max2() of the package base doesn't return an error
# vec <- c(1,3,5,TRUE) ; max2(x = vec, na.rn = TRUE)
}
\seealso{
\code{\link{max}}
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
