% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min2.R
\name{min2}
\alias{min2}
\title{min2}
\usage{
min2(x, na.rm = FALSE, finite = FALSE, safer_check = TRUE)
}
\arguments{
\item{x}{Numeric or logical vector or matrix or numeric table where find the initial values to use.}

\item{na.rm}{Single logical value. Should missing values (NA and NaN) be removed ?}

\item{finite}{Single logical value. Should infinite values (Inf and -Inf) be removed ? Warning: this argument does not remove NA and NaN. Please use the na.rm argument.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
The minimum value of the given arguments.
}
\description{
Find and return the minimum value of a set of data.
}
\examples{
vec <- c(1:3) ; min2(x = vec)

vec <- c(1, 3, 5, TRUE, FALSE) ; min2(x = vec)


# This example returns an error because of the character in the vector
# vec <- c(1,3,5,TRUE,"apple") ; min2(x = vec)
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
