% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_change.R
\name{name_change}
\alias{name_change}
\title{name_change}
\usage{
name_change(data1, data2, added.string = "_modif")
}
\arguments{
\item{data1}{Vector of character strings to check and modify.}

\item{data2}{Reference vector of character strings.}

\item{added.string}{Single character string added at the end of the modified string in data1 if present in data2.}
}
\value{
A list containing :

- $data: the modified data1 (in the same order as in the initial data1).

- $ini: the initial elements before modification. NULL if no modification.

- $post: the modified elements in the same order as in ini. NULL if no modification.
}
\description{
This function allow to check if a vector of character strings, like column names of a data frame, has elements present in another vector (vector of reserved words or column names of another data frame before merging).
}
\examples{
obs1 <- c("A", "B", "C", "D") ; 
obs2 <- c("A", "C") ; 
name_change(obs1, obs2)


obs1 <- c("A", "B", "C", "C_modif1", "D") ; 
obs2 <- c("A", "A_modif1", "C") ; 
name_change(obs1, obs2) 
# the function checks that the new names are neither in obs1 nor in obs2 (increment the number after the added string)
}
