% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail2.R
\name{tail2}
\alias{tail2}
\title{tail2}
\usage{
tail2(data1, n = 6, side = "l", safer_check = TRUE)
}
\arguments{
\item{data1}{Any object but more dedicated for matrix, data frame or table.}

\item{n}{As in utils::tail() but for for matrix, data frame or table, number of dimension to print (10 means 10 rows and columns).}

\item{side}{Either "l" or "r" for the left or right side of the 2D object (only for matrix, data frame or table).}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Must be set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
The tail.
}
\description{
As utils::tail() but display the left or right head of big 2D objects.
}
\details{
Warning: other arguments of tail() not used.
}
\examples{
obs1 = matrix(1:30, ncol = 5, dimnames = list(letters[1:6], LETTERS[1:5])) ; 
obs1 ; 
tail2(obs1, 3, "r")
tail2(letters, 3)
}
\seealso{
\code{\link[utils]{tail}}
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
